Commit 2
First sample z to then sample theta1, theta2, and theta3. This way, we don't get big jumps in loglikelihood. This was happening before because we would propose a new group and the algorithm would not give a chance for theta1, theta2, and theta3 to update correspondingly, resulting in terrible loglikelihoods